Move CodeMappingInfo to Decompiler.
